requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
fastapi==0.116.1
fonttools==4.59.0
h11==0.16.0
idna==3.10
joblib==1.5.1
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.3.1
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
scikit-learn==1.7.1
scipy==1.16.0
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.35.0
fastapi
uvicorn
fastapi
uvicorn
scikit-learn
joblib
numpy
pandas


